Fix signed comparision

+++ unittests/chain_tests.cpp
@@ -101,7 +101,7 @@ BOOST_AUTO_TEST_CASE_TEMPLATE( decompressed_size_under_limit, T, testers ) try {
    constexpr size_t num_cf_actions = 129;
    trx.context_free_actions.insert(trx.context_free_actions.end(), num_cf_actions, act);
    // this is a over limit size 4*129*1024 = ~2M < 10M
-   for(int i = 0; i < num_cf_actions; ++i){
+   for(size_t i = 0; i < num_cf_actions; ++i){
       vector<uint32_t> v(1024, 1);
       trx.context_free_data.emplace_back(fc::raw::pack<vector<uint32_t>>(v));
    }
